function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.8.0
psycopg2-binary==2.8.5
python-http-client==3.3.3
sendgrid==6.3.0